Added test coverage 74%

+++ src/users/users.controller.ts
@@ -7,8 +7,10 @@ import {
   Param,
   Delete,
   UseGuards,
+  BadRequestException,
+  NotFoundException,
 } from '@nestjs/common';
-import { AccessGuard } from 'src/auth/guards/at.guard';
+import { AccessGuard } from './../auth/guards/at.guard';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersService } from './users.service';
@@ -19,7 +21,15 @@ export class UsersController {
 
   @Post()
   @UseGuards(AccessGuard)
-  create(@Body() createUserDto: CreateUserDto) {
+  async create(@Body() createUserDto: CreateUserDto) {
+    const existedUser = await this.usersService.findOneByFilter({
+      email: createUserDto.email,
+    });
+
+    if (existedUser) {
+      throw new BadRequestException('User with such email already exists.');
+    }
+
     return this.usersService.create(createUserDto);
   }
 
@@ -29,19 +39,34 @@ export class UsersController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.usersService.findOne(id);
+  async findOne(@Param('id') id: string) {
+    const user = await this.usersService.findOne(id);
+    if (!user) {
+      throw new NotFoundException('No user found with such ID.');
+    }
+    return user;
   }
 
   @Patch(':id')
   @UseGuards(AccessGuard)
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  async update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    const existedUser = await this.usersService.findOne(id);
+
+    if (!existedUser) {
+      throw new NotFoundException("User with such ID doesn't exists.");
+    }
+
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
   @UseGuards(AccessGuard)
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string) {
+    const existedUser = await this.usersService.findOne(id);
+
+    if (!existedUser) {
+      throw new NotFoundException("User with such ID doesn't exists.");
+    }
     return this.usersService.remove(id);
   }
 }
